sharkabc3d: Shark and Ray Abiotic Covariates in 3 Dimensions

Facilitates the analysis of shark and ray habitat in 3D, enabling descriptions of habitat by depth and area. Provides tools for volume calculation, environmental extraction, and spatial overlap analysis between species ranges, fisheries, and depth-stratified environmental data.

See also

Author

Maintainer: Jay Matsushiba hello@jmatsushiba.com (ORCID)